Fixed several compiler warnings and IndexType issues.

+++ scai/dmemo/GridDistribution.cpp
@@ -91,7 +91,7 @@ GridDistribution::GridDistribution( const Grid& globalGrid, const CommunicatorPt
     {
         PartitionId procGrid[2];
 
-        double size = globalGrid.size( 0 ) + globalGrid.size( 1 );
+        double size = static_cast<double>( globalGrid.size( 0 ) + globalGrid.size( 1 ) );
 
         double w1 = globalGrid.size( 0 ) / size;
         double w2 = globalGrid.size( 1 ) / size;
@@ -105,7 +105,7 @@ GridDistribution::GridDistribution( const Grid& globalGrid, const CommunicatorPt
     {
         PartitionId procGrid[3];
 
-        double size = globalGrid.size( 0 ) + globalGrid.size( 1 ) + globalGrid.size( 2 );
+        double size = static_cast<double>( globalGrid.size( 0 ) + globalGrid.size( 1 ) + globalGrid.size( 2 ) );
 
         double w1 = globalGrid.size( 0 ) / size;
         double w2 = globalGrid.size( 1 ) / size;
